--- src/lib/application/admin/productAdmin/bloc/ProductAdmin_Blocs.dart
@@ -1,3 +1,0 @@
-export 'ProductAdmin_Bloc.dart';
-export '../events/ProductAdmin_Event.dart';
-export '../states/ProductAdmin_State.dart';
