--- src/lib/domain/admin/userAdmin/entity/UserAdminModels.dart
@@ -1,1 +1,0 @@
-export 'UserAdmin.dart';
