+++ src/lib/domain/product/entity/ProductAdmin.dart
