--- src/lib/infrastructure/admin/machineryAdmin/repository/MachineryAdmin_Repository.dart
@@ -1,25 +1,0 @@
-import 'package:mitane_frontend/infrastructure/admin/machineryAdmin/data_provider/MachineryAdmin_Data-Provider.dart';
-import 'package:mitane_frontend/domain/admin/machineryAdmin/entity/MachineryAdmin.dart';
-
-
-class MachineryAdminRepository {
-  final MachineryAdminDataProvider dataProvider;
-  MachineryAdminRepository({required this.dataProvider});
-
-  Future<Machinery> create(Machinery machinery) async {
-    return this.dataProvider.create(machinery);
-  }
-
-  Future<Machinery> update(String id, Machinery machinery) async {
-    return this.dataProvider.update(id, machinery);
-  }
-
-  Future<List<Machinery>> fetchAll() async {
-    print("Fetched");
-    return this.dataProvider.fetchAll();
-  }
-
-  Future<void> delete(String id) async {
-    this.dataProvider.delete(id);
-  }
-}
