+++ src/lib/infrastructure/product/repository/ProductAdmin_Repository.dart
@@ -1,0 +1,25 @@
+import 'package:mitane_frontend/infrastructure/product/data_provider/ProductAdmin_Data-Provider.dart';
+import 'package:mitane_frontend/domain/product/entity/ProductAdmin.dart';
+
+
+class ProductAdminRepository {
+  final ProductAdminDataProvider dataProvider;
+  ProductAdminRepository({required this.dataProvider});
+
+  Future<Product> create(Product product) async {
+    return this.dataProvider.create(product);
+  }
+
+  Future<Product> update(String name, Product product) async {
+    return this.dataProvider.update(name, product);
+  }
+
+  Future<List<Product>> fetchAll() async {
+    print("Fetched");
+    return this.dataProvider.fetchAll();
+  }
+
+  Future<void> delete(String name) async {
+    this.dataProvider.delete(name);
+  }
+}
