+++ src/lib/main.dart
@@ -3,10 +3,9 @@ import 'dart:js';
 import 'package:dio/dio.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:mitane_frontend/application/admin/userAdmin/bloc/UserAdmin_Bloc.dart';
-import 'package:mitane_frontend/application/admin/userAdmin/bloc/UserAdmin_Blocs.dart';
-import 'package:mitane_frontend/infrastructure/admin/userAdmin/data_provider/UserAdmin_Data-Provider.dart';
-import 'package:mitane_frontend/infrastructure/admin/userAdmin/repository/UserAdmin_Repository.dart';
+import 'package:mitane_frontend/application/user/bloc/User_Blocs.dart';
+import 'package:mitane_frontend/infrastructure/user/data_provider/UserAdmin_Data-Provider.dart';
+import 'package:mitane_frontend/infrastructure/user/repository/UserAdmin_Repository.dart';
 import 'package:mitane_frontend/application/price/bloc/price_bloc.dart';
 import 'package:mitane_frontend/application/product/bloc/product_bloc.dart';
 import 'package:mitane_frontend/application/auth/bloc/auth_bloc.dart';
