+++ src/lib/route_generator.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:mitane_frontend/domain/admin/userAdmin/entity/UserAdminModels.dart';
+import 'package:mitane_frontend/domain/user/entity/UserModels.dart';
 import 'package:mitane_frontend/presentation/pages/admin/userAdmin/UserAdmin_Add.dart';
 import 'package:mitane_frontend/presentation/pages/admin/userAdmin/UserAdmin_Lists.dart';
 import 'package:mitane_frontend/presentation/pages/agri_inputs/screens/ingredient_page.dart';
